created package for sortable table + fixed styling
