+++ src/components/sortable_table/SortableTable.js
@@ -1,64 +1,15 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
 import Paper from '@mui/material/Paper';
-import { visuallyHidden } from '@mui/utils';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-
-const theme = createTheme({
-  components: {
-    MuiTableSortLabel: {
-      styleOverrides: {
-        root: {
-          textAlign: 'center',
-          color: '#626262',
-          '&.Mui-active': {
-            color: '#626262'
-          }
-        },
-      }
-    },
-    MuiTablePagination: {
-      styleOverrides: {
-        root: {
-          fontFamily: 'Montserrat',
-          color: '#000',
-          fontSize: '14px',
-          fontWeight: '400',
-          padding: '7px 20px 7px 20px'
-        },
-        actions: {
-          marginRight: '20px',
-          marginLeft: '25px'
-        },
-        displayedRows: {
-          fontFamily: 'Montserrat',
-          color: '#626262',
-          fontSize: '12px',
-        }
-      }
-    },
-    MuiTableCell: {
-      styleOverrides: {
-        root: {
-          fontFamily: 'Montserrat',
-          color: '#000',
-          fontSize: '14px',
-          fontWeight: '400',
-          padding: '7px 20px 7px 20px'
-        }
-      }
-    },
-  }
-});
+import { ThemeProvider } from '@mui/material/styles';
+import EnhancedTableHead from './EnhancedTableHead';
+import { theme } from './Themes';
 
 function descendingComparator(a, b, orderBy) {
   if (b[orderBy] < a[orderBy]) {
@@ -87,63 +38,6 @@ function stableSort(array, comparator) {
   });
   return stabilizedThis.map((el) => el[0]);
 }
-
-function EnhancedTableHead(props) {
-  const {
-    order,
-    orderBy,
-    onRequestSort,
-    headCells
-  } = props;
-  const createSortHandler = (property) => (event) => {
-    onRequestSort(event, property);
-  };
-
-  return (
-    <ThemeProvider theme={theme}>
-      <TableHead>
-        <TableRow>
-          {headCells.map((headCell) => (
-            <TableCell
-              key={headCell.id}
-              align={headCell.numeric ? 'right' : 'left'}
-              padding={headCell.disablePadding ? 'none' : 'normal'}
-              sortDirection={orderBy === headCell.id ? order : false}
-              sx={{
-                lineHeight: 'normal',
-                fontWeight: '600',
-                padding: '10px'
-              }}
-            >
-              <TableSortLabel
-                active={orderBy === headCell.id}
-                direction={orderBy === headCell.id ? order : 'asc'}
-                onClick={createSortHandler(headCell.id)}
-                sx={{
-                  fontSize: '16px',
-                  fontWeight: '400'
-                }}
-              >
-                {headCell.label}
-                {orderBy === headCell.id ? (
-                  <Box component='span' sx={visuallyHidden}>
-                    {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                  </Box>
-                ) : null}
-              </TableSortLabel>
-            </TableCell>
-          ))}
-        </TableRow>
-      </TableHead>
-    </ThemeProvider>
-  );
-}
-
-EnhancedTableHead.propTypes = {
-  onRequestSort: PropTypes.func.isRequired,
-  order: PropTypes.oneOf(['asc', 'desc']).isRequired,
-  orderBy: PropTypes.string.isRequired,
-};
 
 export default function SortableTable(props) {
 
@@ -218,15 +112,14 @@ export default function SortableTable(props) {
           strokeWidth: '1px',
           stroke: '#ABABAB',
           filter: 'drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25))'
-        }}
-      >
+        }} >
         <Paper
           sx={{
             width: '100%',
             mb: 2,
             fontFamily: 'Montserrat',
             fontSize: '100'
-          }}>
+          }} >
           <TableContainer>
             <Table>
               <EnhancedTableHead
@@ -234,8 +127,7 @@ export default function SortableTable(props) {
                 orderBy={orderBy}
                 onRequestSort={handleRequestSort}
                 headCells={headCells}
-                sx={{ marginBottom: '15px' }}
-              />
+                sx={{ marginBottom: '15px' }} />
               <TableBody>
                 {visibleRows.map((row, index) => {
                   const isItemSelected = isSelected(row.name);
@@ -247,14 +139,12 @@ export default function SortableTable(props) {
                       onClick={(event) => handleClick(event, row.name)}
                       tabIndex={-1}
                       key={row.datasetIdentifier}
-                      selected={isItemSelected}
-                    >
+                      selected={isItemSelected} >
                       <TableCell
                         component='th'
                         id={labelId}
                         scope='row'
-                        padding='none'
-                      >
+                        padding='none' >
                         {row.darCode}
                       </TableCell>
                       <TableCell align='right'>{row.approvalDate}</TableCell>
@@ -268,8 +158,7 @@ export default function SortableTable(props) {
                   <TableRow
                     style={{
                       height: 53 * emptyRows,
-                    }}
-                  >
+                    }} >
                     <TableCell colSpan={6} />
                   </TableRow>
                 )}
@@ -283,8 +172,7 @@ export default function SortableTable(props) {
             rowsPerPage={rowsPerPage}
             page={page}
             onPageChange={handleChangePage}
-            onRowsPerPageChange={handleChangeRowsPerPage}
-          />
+            onRowsPerPageChange={handleChangeRowsPerPage} />
         </Paper>
       </Box >
     </ThemeProvider>
